Move US#33 show test out of index_spec into show_spec and test model method format_created_at
