+++ spec/features/admin/dashboard/index_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe "Admin Dashboard Index" do
 
     # Then I see a header indicating that I am on the admin dashboard
     expect(page).to have_content("Admin Dashboard Official")
-    save_and_open_page
+    # save_and_open_page
   end
 
   xit "has links to the Admin Merchants Index and Admind Invouces Index" do
