+++ spec/models/invoice_spec.rb
@@ -11,4 +11,11 @@ RSpec.describe Invoice, type: :model do
   describe "enums" do
     it { should define_enum_for(:status).with_values([:"in progress", :completed, :cancelled]) }
   end
+
+  it "tests timestamp to date method" do
+  customer = create(:customer)
+  invoice = customer.invoices.create!(status: "cancelled", created_at: DateTime.new(2024, 1, 7, 15, 11, 43, '+0'))
+
+  expect(invoice.format_created_at).to eq("Sunday, January 07, 2024")
+  end
 end
